* sysdeps/s390/pthread_spin_lock.c (pthread_spin_lock): Use type
	int for variable OLDVAL and correct inline assembler contraint.
	* sysdeps/s390/pthread_spin_trylock.c (pthread_spin_trylock): Use
	type int for variable OLD.

	* sysdeps/s390/tls.h (TLS_MULTIPLE_THREADS_IN_TCB): Define it
	only for s390-32.
	* sysdeps/unix/sysv/linux/s390/s390-64/sysdep-cancel.h
	(SINGLE_THREAD_P): Use global variable __local_multiple_threads
	instead of multiple_threads field in the TCB.

+++ nptl/sysdeps/s390/pthread_spin_trylock.c
@@ -24,7 +24,7 @@ int
 pthread_spin_trylock (lock)
      pthread_spinlock_t *lock;
 {
-  unsigned long int old;
+  int old;
 
   __asm __volatile ("cs %0,%3,%1"
 		    : "=d" (old), "=Q" (*lock)
